Bouton pour l'ajout d'un projet n'était plus affiché. Parce que le fichier a été déplacé. Modification pour que ça fonctionne.
