+++ LLIO2025/src/Components/Projects/ProjectsLeftPane.svelte
@@ -1,74 +1,30 @@
 <script lang="ts">
-  import { goto } from "$app/navigation"
-  import { quintOut } from "svelte/easing";
-  import { slide } from "svelte/transition";
-  import ProjectItem from "./ProjectPaneItem.svelte";
+  import { goto } from '$app/navigation';
+  import { quintOut } from 'svelte/easing';
+  import { slide } from 'svelte/transition';
+  import ProjectItem from './ProjectPaneItem.svelte';
+  import { Plus } from 'lucide-svelte';
+  import ProjectModal from '../Project/ProjectModal.svelte';
+  import type { CreateProject } from '../../Models';
 
   let { projects } = $props();
   let isArchivedVisible = $state(false);
+  let showModal = $state(false);
+  let editToProject = $state(null);
+
+  const handleProjectSubmit = async (projectData: CreateProject) => {
+    //
+  };
+
+  const handleProjectUpdate = async (projectData: CreateProject) => {
+    //
+  };
+
+  function handleNewProject() {
+    editToProject = null;
+    showModal = true;
+  }
 </script>
-
-<div class="dashboard-panel">
-  <!-- En-tête du dashboard -->
-  <div class="dashboard-header">Tableau de bord</div>
-
-  <!-- Contenu du dashboard -->
-  <div class="dashboard-content">
-    <!-- Éléments du dashboard -->
-    <div class="dashboard-item">
-      <div class="inline-flex rounded-md shadow-xs" role="group">
-        <button 
-          onclick={() => goto('./calendar')}
-          type="button" 
-          class="py-2 px-4 text-sm transition-colors font-semibold bg-gray-200 text-gray-900 rounded-l-lg hover:bg-[#014446] hover:text-white cursor-pointer"
-          >
-          Calendrier
-        </button>
-        <button 
-          type="button" 
-          class="px-4 py-2 text-sm transition-colors font-semibold bg-[#014446] text-white rounded-r-lg"
-          >
-          Projets
-        </button>
-      </div>
-    </div>
-
-    <div class="overflow-y-auto max-h-[calc(100vh-150px)]">
-      {#each projects.filter(x => !x.isArchived) as project}
-        <ProjectItem project={project} />
-      {/each}
-
-      <!-- Projets archivés -->
-      {#if projects.some(x => x.isArchived)}
-        <div>
-          <button
-            class="w-full p-4 flex items-center justify-between text-gray-600 hover:bg-gray-50 cursor-pointer"
-            onclick={() => (isArchivedVisible = !isArchivedVisible)}
-          >
-            <span class="font-medium">Projets archivés ({projects.filter(x => x.isArchived).length})</span>
-            <svg
-              class="w-5 h-5 transform transition-transform {isArchivedVisible ? 'rotate-180' : ''}"
-              fill="none"
-              stroke="currentColor"
-              viewBox="0 0 24 24"
-            >
-              <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M19 9l-7 7-7-7"
-              ></path>
-            </svg>
-          </button>
-
-          {#if isArchivedVisible}
-            <div transition:slide={{ duration: 300, easing: quintOut }}>
-              {#each projects.filter(x => x.isArchived) as project}
-                <ProjectItem project={project} />
-              {/each}
-            </div>
-          {/if}
-        </div>
-      {/if}
-    </div>
-  </div>
-</div>
 
 <style>
   .dashboard-panel {
@@ -108,3 +64,92 @@
     background-color: #f5f5f5;
   }
 </style>
+
+<div class="dashboard-panel">
+  <!-- En-tête du dashboard -->
+  <div class="dashboard-header">Tableau de bord</div>
+
+  <!-- Contenu du dashboard -->
+  <div class="dashboard-content">
+    <!-- Éléments du dashboard -->
+    <div class="dashboard-item">
+      <div class="inline-flex rounded-md shadow-xs" role="group">
+        <button
+          onclick={() => goto('./calendar')}
+          type="button"
+          class="py-2 px-4 text-sm transition-colors font-semibold bg-gray-200 text-gray-900 rounded-l-lg hover:bg-[#014446] hover:text-white cursor-pointer"
+        >
+          Calendrier
+        </button>
+        <button
+          type="button"
+          class="px-4 py-2 text-sm transition-colors font-semibold bg-[#014446] text-white rounded-r-lg"
+        >
+          Projets
+        </button>
+      </div>
+
+      <button
+        type="button"
+        onclick={handleNewProject}
+        class="ml-12 px-3 py-2 text-sm transition-colors font-semibold bg-gray-200 text-gray-900 rounded-lg hover:bg-[#014446] hover:text-white cursor-pointer"
+      >
+        <Plus class="h-4 w-4" />
+      </button>
+    </div>
+
+    <div class="overflow-y-auto max-h-[calc(100vh-150px)]">
+      {#each projects.filter((x) => !x.isArchived) as project}
+        <ProjectItem {project} />
+      {/each}
+
+      <!-- Projets archivés -->
+      {#if projects.some((x) => x.isArchived)}
+        <div>
+          <button
+            class="w-full p-4 flex items-center justify-between text-gray-600 hover:bg-gray-50 cursor-pointer"
+            onclick={() => (isArchivedVisible = !isArchivedVisible)}
+          >
+            <span class="font-medium"
+              >Projets archivés ({projects.filter((x) => x.isArchived).length})</span
+            >
+            <svg
+              class="w-5 h-5 transform transition-transform {isArchivedVisible ? 'rotate-180' : ''}"
+              fill="none"
+              stroke="currentColor"
+              viewBox="0 0 24 24"
+            >
+              <path
+                stroke-linecap="round"
+                stroke-linejoin="round"
+                stroke-width="2"
+                d="M19 9l-7 7-7-7"
+              ></path>
+            </svg>
+          </button>
+
+          {#if isArchivedVisible}
+            <div transition:slide={{ duration: 300, easing: quintOut }}>
+              {#each projects.filter((x) => x.isArchived) as project}
+                <ProjectItem {project} />
+              {/each}
+            </div>
+          {/if}
+        </div>
+      {/if}
+    </div>
+  </div>
+</div>
+
+<!-- Modal du projet qui s'affiche par-dessus tout le reste -->
+{#if showModal}
+  <ProjectModal
+    show={showModal}
+    projectToEdit={editToProject}
+    onSubmit={handleProjectSubmit}
+    onUpdate={handleProjectUpdate}
+    onClose={() => {
+      showModal = false;
+    }}
+  />
+{/if}
